data/UserManagement/AkshaUserManagement/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp
azure-functions
azure-data-tables
azure-servicebus
azure-storage-blob
azure-identity
